enable zetachain support
