+++ api/.env.example
@@ -14,7 +14,7 @@ CONTRACT_ABI=[{"anonymous":false,"inputs":[{"indexed":true,"internalType":"bytes
 INTENT_INITIATED_EVENT_ABI=[{"anonymous":false,"inputs":[{"indexed":true,"internalType":"bytes32","name":"intentId","type":"bytes32"},{"indexed":true,"internalType":"address","name":"asset","type":"address"},{"indexed":false,"internalType":"uint256","name":"amount","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"targetChain","type":"uint256"},{"indexed":false,"internalType":"bytes","name":"receiver","type":"bytes"},{"indexed":false,"internalType":"uint256","name":"tip","type":"uint256"},{"indexed":false,"internalType":"uint256","name":"salt","type":"uint256"}],"name":"IntentInitiated","type":"event"}]
 
 # ZetaChain configuration
-ZETA_CHAIN_ID=7001
+ZETA_CHAIN_ID=7000
 ZETA_RPC_URL=wss://zetachain-athens-evm.blockpi.network/v1/rpc/public
 ZETA_ROUTER_ADDR=0x0000000000000000000000000000000000000000
 ZETA_BLOCK_INTERVAL=1
