+++ frontend/package.json
@@ -15,7 +15,7 @@
     "format:check": "prettier --check ."
   },
   "dependencies": {
-    "@rainbow-me/rainbowkit": "^1.3.3",
+    "@rainbow-me/rainbowkit": "1.3.3",
     "@tanstack/react-query": "^4.36.1",
     "ethers": "^6.13.5",
     "next": "14.1.0",
@@ -24,7 +24,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "viem": "^1.21.4",
-    "wagmi": "^1.4.13"
+    "wagmi": "1.4.13"
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^15.2.4",
