+++ frontend/src/components/Web3Provider.tsx
@@ -5,11 +5,6 @@ import "@rainbow-me/rainbowkit/styles.css";
 import { WagmiConfig } from "wagmi";
 import { config, chains } from "../utils/web3Config";
 import { arcadeTheme } from "../utils/rainbowKitTheme";
-import { RpcTest } from "./RpcTest";
-
-if (!process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID) {
-  throw new Error("NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID is not defined");
-}
 
 export function Web3Provider({ children }: { children: React.ReactNode }) {
   return (
@@ -17,10 +12,8 @@ export function Web3Provider({ children }: { children: React.ReactNode }) {
       <RainbowKitProvider
         chains={chains}
         theme={arcadeTheme}
-        coolMode
-        showRecentTransactions={true}
+        modalSize="compact"
       >
-        <RpcTest />
         {children}
       </RainbowKitProvider>
     </WagmiConfig>
