+++ frontend/src/config/index.ts
@@ -1,0 +1,2 @@
+// Export all chain and token configuration from a single place
+export * from "./chainConfig";
