+++ frontend/src/services/contract.ts
@@ -1,7 +1,6 @@
 import { useAccount, usePublicClient, useWalletClient } from "wagmi";
 import { getContract } from "viem";
 import { arbitrum } from "viem/chains";
-import { Abi, GetContractReturnType, PublicClient, WalletClient } from "viem";
 import { Intent as IntentContract } from "../contracts/Intent";
 
 export function useContractService() {
@@ -13,7 +12,7 @@ export function useContractService() {
     if (!publicClient || !walletClient) return null;
 
     return getContract({
-      address: IntentContract.address[42161] as `0x${string}`, // Using Arbitrum address
+      address: IntentContract.address[arbitrum.id] as `0x${string}`, // Using Arbitrum address
       abi: IntentContract.abi,
       publicClient,
       walletClient,
