+++ frontend/src/utils/rainbowKitTheme.ts
@@ -3,7 +3,7 @@ import { darkTheme } from "@rainbow-me/rainbowkit";
 export const arcadeTheme = darkTheme({
   accentColor: "#FFD700",
   accentColorForeground: "#000000",
-  borderRadius: "medium",
+  borderRadius: "small",
   fontStack: "system",
   overlayBlur: "small",
 });
