+++ frontend/src/utils/web3Config.ts
@@ -7,46 +7,50 @@ import {
 import { configureChains, createConfig } from "wagmi";
 import { alchemyProvider } from "wagmi/providers/alchemy";
 import { publicProvider } from "wagmi/providers/public";
-import { getCustomChains } from "@/config/chains";
+import { getCustomChains } from "@/config/chainConfig";
 
-if (!process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID) {
-  throw new Error("Missing NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID");
-}
-
-if (!process.env.NEXT_PUBLIC_ALCHEMY_ID) {
-  throw new Error("Missing NEXT_PUBLIC_ALCHEMY_ID");
-}
-
+// Check for required environment variables
 const projectId = process.env.NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID;
 const alchemyId = process.env.NEXT_PUBLIC_ALCHEMY_ID;
 
-console.log("Configuring chains with Alchemy ID:", alchemyId);
+if (!projectId) {
+  console.error("Missing NEXT_PUBLIC_WALLETCONNECT_PROJECT_ID environment variable");
+}
+
+if (!alchemyId) {
+  console.error("Missing NEXT_PUBLIC_ALCHEMY_ID environment variable");
+}
+
+console.log("Configuring chains with Alchemy ID:", alchemyId ? "✓ Available" : "✗ Missing");
 
 // Get custom chain configurations from our centralized config
-const customChains = getCustomChains(alchemyId);
+const customChains = getCustomChains(alchemyId || "demo");
 
-const { chains, publicClient } = configureChains(customChains, [
-  alchemyProvider({ apiKey: alchemyId }),
-  publicProvider(), // Fallback to public provider
-]);
+// Configure chains with appropriate providers
+const providers = [];
+if (alchemyId) {
+  providers.push(alchemyProvider({ apiKey: alchemyId }));
+}
+providers.push(publicProvider());
 
+const { chains, publicClient } = configureChains(customChains, providers);
+
+// Debug logging for chain configuration
 console.log(
-  "Configured chains:",
-  chains.map((chain) => ({
-    id: chain.id,
-    name: chain.name,
-    rpcUrls: chain.rpcUrls,
-  })),
+  "Chains configured:",
+  chains.map((chain) => chain.name).join(", ")
 );
 
+// Configure wallet options
 const { wallets } = getDefaultWallets({
   appName: "Speedrun",
-  projectId,
+  projectId: projectId || "",
   chains,
 });
 
 const connectors = connectorsForWallets([...wallets]);
 
+// Create wagmi config
 const config = createConfig({
   autoConnect: true,
   connectors,
